ARM: move VFP init to an earlier boot stage
In order to use the NEON unit in the kernel, we should
initialize it a bit earlier in the boot process so NEON users
that like to do a quick benchmark at load time (like the
xor_blocks or RAID-6 code) find the NEON/VFP unit already
enabled.

Replaced late_initcall() with core_initcall().

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Acked-by: Nicolas Pitre <nico@linaro.org>

https://github.com/Team-CodeCraft/android_kernel_htc_msm8974/commit/83116b24d1d6a1c6f7697c2b3a31f621f5e9f0da